requirements.txt
asgiref==3.4.1
Django==3.2.7
django-crispy-forms==1.12.0
django-localflavor==3.1
django-easy-select2==1.5.6
python-stdnum==1.18
pytz==2021.1
sqlparse==0.4.2
validate-docbr==1.10.0
djangorestframework==3.12.4
psycopg2-binary>=2.9.1

